Downloads/RoverOps-Project (1) 2/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai>=0.0.2
langchain-community==0.0.20
openai>=1.6.1
python-dotenv==1.0.0
httpx==0.25.0
pydantic>=2.9.0
typing-extensions>=4.12.2

